+++ src/histogram/bins/mod.rs
@@ -1,0 +1,5 @@
+pub use self::one_dim::{Bin1d, Bins1d};
+pub use self::n_dim::{BinNd, BinsNd};
+
+mod one_dim;
+mod n_dim;
